+++ contracts/src/gateway/HomeGateway.sol
@@ -29,7 +29,7 @@ contract HomeGateway is IHomeGateway, UUPSProxiable, Initializable {
     // *             Storage               * //
     // ************************************* //
 
-    address public governor;
+    address public owner;
     IArbitratorV2 public arbitrator;
     IVeaInbox public veaInbox;
     uint256 public override foreignChainID;
@@ -43,9 +43,9 @@ contract HomeGateway is IHomeGateway, UUPSProxiable, Initializable {
     // *        Function Modifiers         * //
     // ************************************* //
 
-    /// @dev Requires that the sender is the governor.
-    modifier onlyByGovernor() {
-        if (governor != msg.sender) revert GovernorOnly();
+    /// @dev Requires that the sender is the owner.
+    modifier onlyByOwner() {
+        if (owner != msg.sender) revert OwnerOnly();
         _;
     }
 
@@ -59,21 +59,21 @@ contract HomeGateway is IHomeGateway, UUPSProxiable, Initializable {
     }
 
     /// @dev Constructs the `PolicyRegistry` contract.
-    /// @param _governor The governor's address.
+    /// @param _owner The owner's address.
     /// @param _arbitrator The address of the arbitrator.
     /// @param _veaInbox The address of the vea inbox.
     /// @param _foreignChainID The ID of the foreign chain.
     /// @param _foreignGateway The address of the foreign gateway.
     /// @param _feeToken The address of the fee token.
     function initialize(
-        address _governor,
+        address _owner,
         IArbitratorV2 _arbitrator,
         IVeaInbox _veaInbox,
         uint256 _foreignChainID,
         address _foreignGateway,
         IERC20 _feeToken
     ) external reinitializer(1) {
-        governor = _governor;
+        owner = _owner;
         arbitrator = _arbitrator;
         veaInbox = _veaInbox;
         foreignChainID = _foreignChainID;
@@ -87,39 +87,39 @@ contract HomeGateway is IHomeGateway, UUPSProxiable, Initializable {
 
     /**
      * @dev Access Control to perform implementation upgrades (UUPS Proxiable)
-     * @dev Only the governor can perform upgrades (`onlyByGovernor`)
+     * @dev Only the owner can perform upgrades (`onlyByOwner`)
      */
-    function _authorizeUpgrade(address) internal view override onlyByGovernor {
+    function _authorizeUpgrade(address) internal view override onlyByOwner {
         // NOP
     }
 
-    /// @dev Changes the governor.
-    /// @param _governor The address of the new governor.
-    function changeGovernor(address _governor) external onlyByGovernor {
-        governor = _governor;
+    /// @dev Changes the owner.
+    /// @param _owner The address of the new owner.
+    function changeOwner(address _owner) external onlyByOwner {
+        owner = _owner;
     }
 
     /// @dev Changes the arbitrator.
     /// @param _arbitrator The address of the new arbitrator.
-    function changeArbitrator(IArbitratorV2 _arbitrator) external onlyByGovernor {
+    function changeArbitrator(IArbitratorV2 _arbitrator) external onlyByOwner {
         arbitrator = _arbitrator;
     }
 
     /// @dev Changes the vea inbox, useful to increase the claim deposit.
     /// @param _veaInbox The address of the new vea inbox.
-    function changeVea(IVeaInbox _veaInbox) external onlyByGovernor {
+    function changeVea(IVeaInbox _veaInbox) external onlyByOwner {
         veaInbox = _veaInbox;
     }
 
     /// @dev Changes the foreign gateway.
     /// @param _foreignGateway The address of the new foreign gateway.
-    function changeForeignGateway(address _foreignGateway) external onlyByGovernor {
+    function changeForeignGateway(address _foreignGateway) external onlyByOwner {
         foreignGateway = _foreignGateway;
     }
 
     /// @dev Changes the fee token.
     /// @param _feeToken The address of the new fee token.
-    function changeFeeToken(IERC20 _feeToken) external onlyByGovernor {
+    function changeFeeToken(IERC20 _feeToken) external onlyByOwner {
         feeToken = _feeToken;
     }
 
@@ -239,7 +239,7 @@ contract HomeGateway is IHomeGateway, UUPSProxiable, Initializable {
     // *              Errors               * //
     // ************************************* //
 
-    error GovernorOnly();
+    error OwnerOnly();
     error ArbitratorOnly();
     error FeesPaidInERC20Only();
     error FeesPaidInNativeCurrencyOnly();